+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Release 2.0.0 (2021-05-)
 - Now when you have walls selected, you can hold the Ctrl Key while clicking a wall type in the Scene Controls to instantly change the selected walls to that wall type.
+- Fixed Colour Picker so it no longer activates when you hit Enter in a config form.
 
 ## Release v1.1.0 (2021-04-18)
 - Fixed on-load crash when AltGridSnap is disabled.